feat: /products/:id now can display discounted value
